+++ tests/common2.sh
@@ -22,6 +22,10 @@ fi
 				if ${TEST} "$4" = "-o" ]; then
 					return 1 # false
 				fi
+			elif ${TEST} "$3" = "/bases/base-test" ]; then
+				return 1 # false
+			elif ${TEST} "$3" = "/bases/base-test-nodset" ]; then
+				return 1 # false
 			else
 				return 0 # true
 			fi
@@ -51,6 +55,9 @@ _zfs_is_dataset()
 	elif ${TEST} "$1" = "/jails/pot-test-noconf" ]; then
 		return 0 # true
 	fi
+	if ${TEST} "$1" = "/bases/base-test" ]; then
+		return 0
+	fi
 	return 1 # false
 }
 
@@ -69,6 +76,21 @@ test_is_pot()
 	assertEquals "3" "$?"
 
 	_is_pot pot-test
+	assertEquals "0" "$?"
+}
+
+test_is_base()
+{
+	_is_base
+	assertEquals "1" "$?"
+
+	_is_base nobase
+	assertEquals "1" "$?"
+
+	_is_base base-test-nodset
+	assertEquals "2" "$?"
+
+	_is_base base-test
 	assertEquals "0" "$?"
 }