+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- create-dns : new command, to create the dns pot
+- create : add the option -d to choose the type of dns (inherit or pot)
+- start : add the support to dns types
+
 ### Fixed
 - rename : remove a misleading error message
 - start : proper stop, if one mount fails